uitwerking/requirements.txt
Flask==1.0.2
Jinja2>=2.10.1
Werkzeug==0.14.1
mysql-connector-python==8.0.11
biopython==1.73